+++ x/poolmanager/keeper.go
@@ -98,14 +98,40 @@ func (k Keeper) InitGenesis(ctx sdk.Context, genState *types.GenesisState) {
 	for _, poolRoute := range genState.PoolRoutes {
 		k.SetPoolRoute(ctx, poolRoute.PoolId, poolRoute.PoolType)
 	}
+
+	// We track taker fees generated in the module's KVStore.
+	// If the values were exported, we set them here.
+	// If the values were not exported, we initialize the tracker to zero and set the accounting height to the current height.
+	if !genState.TakerFeesTracker.TakerFeesToStakers.Empty() {
+		k.SetTakerFeeTrackerForStakers(ctx, genState.TakerFeesTracker.TakerFeesToStakers)
+	} else {
+		k.SetTakerFeeTrackerForStakers(ctx, sdk.NewCoins())
+	}
+	if !genState.TakerFeesTracker.TakerFeesToCommunityPool.Empty() {
+		k.SetTakerFeeTrackerForCommunityPool(ctx, genState.TakerFeesTracker.TakerFeesToCommunityPool)
+	} else {
+		k.SetTakerFeeTrackerForCommunityPool(ctx, sdk.NewCoins())
+	}
+	if genState.TakerFeesTracker.HeightAccountingStartsFrom != 0 {
+		k.SetTakerFeeTrackerStartHeight(ctx, genState.TakerFeesTracker.HeightAccountingStartsFrom)
+	} else {
+		k.SetTakerFeeTrackerStartHeight(ctx, ctx.BlockHeight())
+	}
 }
 
 // ExportGenesis returns the poolmanager module's exported genesis.
 func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
+	// Export KVStore values to the genesis state so they can be imported in init genesis.
+	takerFeesTracker := types.TakerFeesTracker{
+		TakerFeesToStakers:         k.GetTakerFeeTrackerForStakers(ctx),
+		TakerFeesToCommunityPool:   k.GetTakerFeeTrackerForCommunityPool(ctx),
+		HeightAccountingStartsFrom: k.GetTakerFeeTrackerStartHeight(ctx),
+	}
 	return &types.GenesisState{
-		Params:     k.GetParams(ctx),
-		NextPoolId: k.GetNextPoolId(ctx),
-		PoolRoutes: k.getAllPoolRoutes(ctx),
+		Params:           k.GetParams(ctx),
+		NextPoolId:       k.GetNextPoolId(ctx),
+		PoolRoutes:       k.getAllPoolRoutes(ctx),
+		TakerFeesTracker: &takerFeesTracker,
 	}
 }