+++ x/txfees/keeper/feedecorator.go
@@ -265,5 +265,7 @@ func DeductFees(txFeesKeeper types.TxFeesKeeper, bankKeeper types.BankKeeper, ct
 		}
 	}
 
+	txFeesKeeper.IncreaseTxFeesTracker(ctx, fees[0])
+
 	return nil
 }